requirements-dev.txt
hiredis==0.2.0
django-nose==1.4.4
nose==1.3.6
msgpack-python==0.4.6
pyyaml==5.1
mock
fakeredis